Fixes wrong parse exception type being caught by TranquilityEventWriter
Avro extensions from the Druid codebase throw a ParseException
from the io.druid.java.util.common.parsers package, the Tranquility
code is not catching this as it is expecting the ParseException to
be from the com.metamx.common.parsers package.

This means that broken messages will halt processing regardless of the
reportParseExceptions setting as the exception goes uncaught.

Fixed this to use the ParseException from io.druid.java.util.common.parsers
package